+++ planner/core/BUILD.bazel
@@ -177,7 +177,6 @@ go_test(
         "indexmerge_test.go",
         "integration_partition_test.go",
         "integration_test.go",
-        "logical_plan_test.go",
         "logical_plan_trace_test.go",
         "logical_plans_test.go",
         "main_test.go",
@@ -248,6 +247,7 @@ go_test(
         "//testkit/ddlhelper",
         "//testkit/external",
         "//testkit/testdata",
+        "//testkit/testmain",
         "//testkit/testsetup",
         "//testkit/testutil",
         "//types",
